requirements.txt
google-generativeai
python-dotenv
flask
flask-socketio
edge-tts
pygame
pillow
sounddevice
SpeechRecognition
PyMuPDF
numpy
pyautogui
pywinauto
pygetwindow
vosk
pyaudio
playwright
beautifulsoup4
duckduckgo-search
psutil
watchdog
opencv-python
pytesseract
pyttsx3
mediapipe